2019.11.2branch1

+++ app/Http/Controllers/PostsController.php
@@ -4,7 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Post;
-
+use App\Http\Controllers\Controller;
 class PostsController extends Controller
 {
     public function index()
@@ -16,68 +16,70 @@ class PostsController extends Controller
     public function bbs()
     {
       $posts = Post::with(['comments'])->orderBy('created_at', 'desc')->paginate(10);
-
       return view('posts.bbs',['posts'=>$posts]);
     }
 
-public function create()
-{
+    public function create()
+    {
     return view('posts.create');
-}
+    }
 
-public function store(Request $request)
-{
+    public function store(Request $request)
+    {
     $params = $request->validate([
         'title' => 'required|max:50',
         'body' => 'required|max:2000',
     ]);
 
-    Post::create($params);
+        auth()->user()->posts()->create($params);
+        //Post::create($params);
 
     return redirect()->route('bbsTop');
-}
+    }
 
-public function show($post_id)
-{
-  $post=Post::findOrFail($post_id);
+    public function show($post_id)
+    {
+      $post=Post::findOrFail($post_id);
 
-  return view('posts.show',[
-    'post'=>$post,
-  ]);
-}
+      return view('posts.show',[
+        'post'=>$post,
+      ]);
+    }
 
-public function edit($post_id)
-{
-  $post=Post::findOrFail($post_id);
+    public function edit($post_id)
+    {
 
-  return view('posts.edit',[
-    'post'=>$post,
-  ]);
-}
+      $post=Post::findOrFail($post_id);
+      $this->authorize('edit', $post);
+      return view('posts.edit', [
+        'post' => $post,
+    ]);
+    }
 
-public function update($post_id,Request $request)
-{
-  $params=$request->validate([
+    public function update($post_id,Request $request)
+    {
+      $params=$request->validate([
     'title'=>'required|max:50',
     'body'=>'required|max:2000',
-  ]);
-$post=Post::findOrFail($post_id);
-$post->fill($params)->save();
+    ]);
+    $post=Post::findOrFail($post_id);
+    $this->authorize('update', $post);
+    $post->fill($params)->save();
+    return redirect()->route('posts.show',['post'=>$post]);
+    }
 
-return redirect()->route('posts.show',['post'=>$post]);
-}
+    public function destroy($post_id)
+    {
+      $post=Post::findOrFail($post_id);
+      $this->authorize('edit', $post);
 
-public function destroy($post_id)
-{
-  $post=Post::findOrFail($post_id);
-
-  \DB::transaction(function() use ($post){
+      \DB::transaction(function() use ($post){
     $post->comments()->delete();
     $post->delete();
-  });
+    });
 
-  return redirect()->route('bbsTop');
-}
+    return redirect()->route('bbsTop');
+    }
 
 
-}
+    }
